Update YouTubeAlchemy.js
bug fixes and improvements. new features: added a button to copy the transcript to the clipboard, allow changing the label text for button one (NotebookLM) and button two (ChatGPT), which also updates the hover text for those icons. to change the label text, enter "Label | domain.com" in the respective URL fields.